move str_dtype to datatypes.py and change import location in other files
