+++ pyccel/parser/semantic.py
@@ -59,12 +59,10 @@ from pyccel.ast.core import Product, FunctionCall
 
 from pyccel.ast.functionalexpr import FunctionalSum, FunctionalMax, FunctionalMin
 
-from pyccel.ast.datatypes import NativeRange
+from pyccel.ast.datatypes import NativeRange, str_dtype
 from pyccel.ast.datatypes import NativeSymbol
 from pyccel.ast.datatypes import DataTypeFactory
 from pyccel.ast.datatypes import NativeInteger, NativeBool, NativeReal, NativeString, NativeGeneric, NativeComplex
-
-from pyccel.ast.type_inference  import str_dtype
 
 from pyccel.ast.numbers import BooleanTrue, BooleanFalse
 from pyccel.ast.numbers import Integer, Float
